+++ .github/actions/pnpm-install/action.yml
@@ -17,5 +17,7 @@ runs:
       with:
         node-version-file: .nvmrc
     - uses: pnpm/action-setup@v4
+    - run: pnpm --version
+      shell: bash
     - run: pnpm i --frozen-lockfile
       shell: bash
