+++ packages/vuetify/src/globals.d.ts
@@ -4,7 +4,6 @@ import 'vue/jsx'
 import type { ComponentInjectOptions, ComponentOptionsMixin, EmitsOptions, SlotsType } from 'vue'
 import type { ComputedOptions, Events, MethodOptions, VNode } from 'vue'
 import type { TouchStoredHandlers } from './directives/touch'
-import type { CustomCommands } from '../test/setup/browser-commands'
 
 declare global {
   interface HTMLCollection {
@@ -169,18 +168,4 @@ declare module 'vue' {
   }
 
   export interface CSSProperties extends CustomProperties {}
-}
-
-interface CustomMatchers<R = unknown> {
-  toHaveBeenTipped: () => R
-  toHaveBeenWarned: () => R
-}
-
-declare module 'vitest' {
-  interface Assertion<T = any> extends CustomMatchers<T> {}
-  interface AsymmetricMatchersContaining extends CustomMatchers {}
-}
-
-declare module '@vitest/browser/context' {
-  interface BrowserCommands extends CustomCommands {}
 }
