+++ packages/vuetify/test/globals.d.ts
@@ -1,0 +1,15 @@
+import type { CustomCommands } from './setup/browser-commands.ts'
+
+interface CustomMatchers<R = unknown> {
+  toHaveBeenTipped: () => R
+  toHaveBeenWarned: () => R
+}
+
+declare module 'vitest' {
+  interface Assertion<T = any> extends CustomMatchers<T> {}
+  interface AsymmetricMatchersContaining extends CustomMatchers {}
+}
+
+declare module '@vitest/browser/context' {
+  interface BrowserCommands extends CustomCommands {}
+}
