+++ packages/vuetify/tsconfig.dist.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["dev", "**/*.spec.*", "cypress"]
+  "exclude": ["dev", "**/*.spec.*", "test"]
 }
